requirements.txt
fastapi
uvicorn
openai
langchain
faiss-cpu
pydantic
beautifulsoup4
requests
tqdm
